+++ feeluown/app/__init__.py
@@ -1,0 +1,6 @@
+from .app import App
+
+
+__all__ = (
+    'App',
+)
