+++ feeluown/app/gui_app.py
@@ -1,0 +1,83 @@
+import os
+
+from PyQt5.QtCore import Qt, QDir
+from PyQt5.QtWidgets import QApplication, QWidget
+
+from feeluown.gui.browser import Browser
+from feeluown.gui.hotkey import HotkeyManager
+from feeluown.gui.image import ImgManager
+from feeluown.gui.theme import ThemeManager
+from feeluown.gui.tips import TipsManager
+from feeluown.gui.ui import Ui
+from feeluown.gui.tray import Tray
+from feeluown.uimodels.provider import ProviderUiManager
+from feeluown.uimodels.playlist import PlaylistUiManager
+from feeluown.uimodels.my_music import MyMusicUiManager
+from feeluown.uimodels.collection import CollectionUiManager
+
+from feeluown.collection import CollectionManager
+
+from .app import App
+
+
+class GuiApp(App, QWidget):
+    def __init__(self, config):
+        App.__init__(config)
+        QWidget.__init__(self)
+
+        pkg_root_dir = os.path.dirname(__file__)
+        icons_dir = os.path.join(pkg_root_dir, 'icons')
+        QDir.addSearchPath('icons', icons_dir)
+
+        self.setObjectName('app')
+
+        # GUI 的一些辅助管理模块
+        self.coll_mgr = CollectionManager(self)
+        self.theme_mgr = ThemeManager(self)
+        self.tips_mgr = TipsManager(self)
+        self.hotkey_mgr = HotkeyManager(self)
+        self.img_mgr = ImgManager(self)
+
+        # GUI 组件的数据管理模块
+        self.pvd_uimgr = ProviderUiManager(self)
+        self.pl_uimgr = PlaylistUiManager(self)
+        self.mymusic_uimgr = MyMusicUiManager(self)
+        self.coll_uimgr = CollectionUiManager(self)
+
+        self.browser = Browser(self)
+        self.ui = Ui(self)
+        if self.config.ENABLE_TRAY:
+            self.tray = Tray(self)
+        #self.show_msg = self.ui.toolbar.status_line.get_item('notify').widget.show_msg
+
+    def initialize(self):
+        self.hotkey_mgr.initialize()
+        self.theme_mgr.initialize()
+        if self.config.ENABLE_TRAY:
+            self.tray.initialize()
+            self.tray.show()
+        self.tips_mgr.show_random_tip()
+        self.coll_uimgr.initialize()
+        self.browser.initialize()
+        self.show()
+        QApplication.instance().aboutToQuit.connect(self.about_to_exit)
+
+    def closeEvent(self, _):
+        if not self.config.ENABLE_TRAY:
+            self.exit()
+
+    def mouseReleaseEvent(self, e):
+        if not self.rect().contains(e.pos()):
+            return
+        if e.button() == Qt.BackButton:
+            self.browser.back()
+        elif e.button() == Qt.ForwardButton:
+            self.browser.forward()
+
+    def exit_player(self):
+        # Destroy GL context or mpv renderer
+        self.ui.mpv_widget.shutdown()
+        super().exit_player()
+
+    def exit(self):
+        QApplication.exit()
