+++ feeluown/entry_points/run.py
@@ -9,24 +9,18 @@ except ImportError:
     pass
 
 # pylint: disable=wrong-import-position
-import os
-import warnings
-import signal
-from feeluown.app import create_app, init_app, forevermain
-from feeluown.cli import oncemain
-from feeluown.fuoexec import fuoexec_load_rcfile, fuoexec_init
-from feeluown.utils import aio
 from feeluown.utils.utils import is_port_inuse  # noqa: E402
-from feeluown.utils.dispatch import Signal  # noqa: E402
-from .base import ensure_dirs, setup_config, setup_logger  # noqa: E402
-from .base import create_config, setup_argparse  # noqa: E402
+from .base import setup_argparse  # noqa: E402
 from .run_cli import run_cli  # noqa: E402
+from .run_app import run_app
 
 
 def run():
-    """feeluown entry point"""
+    """feeluown entry point.
+    """
 
     args = setup_argparse().parse_args()
+
     if args.cmd is not None:  # Only need to run some commands.
         if args.cmd == 'genicon':
             return run_cli(args)
@@ -37,7 +31,6 @@ def run():
         if is_port_inuse(2333):
             return run_cli(args)
 
-        forever = False
         # If daemon is not started, (some) commands can be meaningless,
         # such as `status`, `toggle`, `next`, etc. However, some other
         # commands can still be usefule. For instance, when people
@@ -50,76 +43,9 @@ def run():
         cmds = ('show', 'play', 'search')
         if args.cmd not in cmds:
             raise SystemExit("can't connect to fuo daemon.")
+
+        adhoc = False
     else:
-        forever = True
+        adhoc = True
 
-    config = create_config()
-
-    if forever is True:
-        # ensure requiresments, raise SystemExit if failed
-        os.chdir(os.path.join(os.path.dirname(__file__), '../..'))
-        ensure_dirs()
-    else:
-        # ignore all warnings since it will pollute the output
-        warnings.filterwarnings("ignore")
-
-    # In an ideal world, users are capable to do anything in rcfile,
-    # including monkeypatch, so we should load rcfile as early as possible
-    fuoexec_load_rcfile(config)
-
-    # Extract config items from args and setup config object,
-    # since then, no more `args` object, only `config`.
-    setup_config(args, config)
-
-    if forever is True:
-        setup_logger(config)
-
-    # create app instance with config
-    #
-    # nothing should happen except all objects are created
-    app = create_app(config)
-
-    async def run_app():
-        Signal.setup_aio_support()
-
-        # do fuoexec initialization before app inits
-        fuoexec_init(app)
-
-        # initialize app with config
-        #
-        # all objects can do initialization here. some objects may emit signal,
-        # some objects may connect with others signals.
-        init_app(app)
-
-        if forever is True:
-            app.load_state()
-
-        app.initialized.emit(app)
-
-        if forever is True:
-            await forevermain(app)
-        else:
-            await oncemain(app, args)
-
-
-    # App can exit in several ways
-    #
-    # GUI mode
-    # 1. user clicks the tray icon exit button, which triggers QApplication.quit.
-    # 2. SIGTERM is received.
-    # 3. QApplication.quit is called. (For example, User press CMD-Q on macOS.)
-    #
-    # Daemon mode
-    # 1. Ctrl-C
-    # 2. SIGTERM
-    def handle_signal(signum, _):
-        if signum == signal.SIGTERM:
-            app.exit()
-
-    signal.signal(signal.SIGTERM, handle_signal)
-
-    try:
-        aio.run(run_app())
-    except KeyboardInterrupt:
-        if forever is True:
-            app.exit()
+    return run_app(args, adhoc=adhoc)
